chore: refactor scroll-utils together into utils
