+++ src/flashbar/utils.ts
@@ -123,3 +123,11 @@ export const counterTypes: {
   { type: 'info', labelName: 'infoIconAriaLabel', iconName: 'status-info' },
   { type: 'progress', labelName: 'inProgressIconAriaLabel', iconName: 'status-in-progress' },
 ];
+
+export function isElementTopBeyondViewport(element: HTMLElement) {
+  return element.getBoundingClientRect().top < 0;
+}
+
+export function isKeyboardInteraction(isFocusVisible: Record<string, boolean | unknown>) {
+  return !!Object.keys(isFocusVisible).length;
+}
